# docs/swagger.yaml
basePath: /
definitions:
  models.ActiveUser:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.ActiveUserResponse'
    type: object
  models.ActiveUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.Board:
    properties:
      destination_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  models.Login:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  models.UpdateBoard:
    properties:
      name:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      team_id:
        items:
          type: string
        type: array
    type: object
  utils.AppError:
    properties:
      code:
        description: HTTP status code
        type: integer
      message:
        description: error message
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A simple task tracker API built with Go and MongoDB
  title: Task Tracker API
  version: "1.0"
paths:
  /boards:
    get:
      description: Retrieve all boards
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Board'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Get all boards
      tags:
      - boards
  /boards/{id}:
    delete:
      description: Delete a board by ID
      parameters:
      - description: Board ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Delete board
      tags:
      - boards
    get:
      description: Retrieve a single board by ID
      parameters:
      - description: Board ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Board'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Get board
      tags:
      - boards
    put:
      consumes:
      - application/json
      description: Update a board by ID
      parameters:
      - description: Board ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated board payload
        in: body
        name: board
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBoard'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Update board
      tags:
      - boards
  /boards/by-destination:
    get:
      description: Retrieve boards filtered by destination ID
      parameters:
      - description: Destination ID
        in: query
        name: destination_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Board'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Get boards by destination
      tags:
      - boards
...
  /destinations/{id}/boards:
    post:
      consumes:
      - application/json
      description: Create a new board under a destination
      parameters:
      - description: Destination ID
        in: path
        name: id
        required: true
        type: string
      - description: Board payload
        in: body
        name: board
        required: true
        schema:
          $ref: '#/definitions/models.Board'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Create board
      tags:
      - boards
  /login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ActiveUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Login
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: User registration payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Account created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.AppError'
      summary: Create account
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"